pyproject.toml
[project]
name = "doc-extract-env"
version = "0.1.0"
description = "Local assessment environment for structured extraction from documents with OCR and layout noise."
readme = "README.md"
requires-python = "==3.12.2"
dependencies = [
  "Faker==40.13.0",
  "jsonschema==4.26.0",
  "numpy==2.4.4",
  "pandas==3.0.2",
  "Pillow==12.2.0",
  "pydantic==2.13.0",
  "python-dateutil==2.9.0.post0",
  "rapidfuzz==3.14.5",
]

[tool.uv]
package = false
